Language/Python/Crawler/requirement.txt
request
requests
requests-html
html
aiohttp
beautifulsoup4
csv
pyocr
pillow
pytesser3
lxml
html5lib
xlrd
xlwt
python-docx
sqlalchemy
pymysql
pymongo

